Refactor: SignIn and SignUp
