+++ src/Profile/SignUp/SignUp.module.css
@@ -1,0 +1,3 @@
+.SignUpFrame {
+    
+}
