+++ src/Profile/Signin/SignInButton.module.css
@@ -12,6 +12,7 @@
   height: 2.375rem;
   position: relative;
   cursor: pointer;
+  margin-top: 0.8rem;
   margin-bottom: 10px;
 
   color: rgba(0, 0, 0, 0.46);
