+++ src/Profile/Signin/SignInInputField.jsx
@@ -18,7 +18,6 @@ function SignInInputField({
           onChange={onChange}
           type={type}
           placeholder={errorMessage ? `⚠ ${errorMessage}` : label} // placeholder도 prop으로부터 받은 label을 사용합니다.
-          // aria-describedby={error ? "error-msg" : undefined}
         />
       </div>
       <div className={styles.nameLabel}>{label}</div>
