+++ src/Profile/Signin/SignInMain.module.css
@@ -1,0 +1,77 @@
+.modal {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.6);
+}
+
+.signInTitle {
+  color: rgba(0, 0, 0, 0.85);
+  font-family: "Alef-Regular", sans-serif;
+  font-size: 2rem;
+  font-weight: 400;
+  margin-bottom: 0.4rem;
+  position: relative;
+  text-align: center;
+}
+
+.signInFrame {
+  background: white;
+  border-radius: 0.625rem;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0 0 0.375rem rgba(0, 0, 0, 0.25);
+  width: 480px;
+  min-height: 621px;
+  background-color: white;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  justify-content: center;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+.MainFrame {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  position: relative;
+}
+
+.close {
+  flex-shrink: 0;
+  width: 2rem;
+  height: 2rem;
+  position: absolute;
+  right: 1.5rem;
+  top: 1.3125rem;
+  overflow: visible;
+  cursor: pointer;
+}
+
+/* SwitchText */
+.SwitchFrame {
+  display: flex;
+  flex-direction: row;
+  position: relative;
+}
+
+.AskSwitch {
+  color: rgba(0, 0, 0, 0.8);
+  font-family: "Alef-Regular", sans-serif;
+  font-size: 1rem;
+  font-weight: 400;
+  padding-right: 0.5rem;
+}
+
+.signUpText {
+  color: rgba(13, 69, 179, 0.96);
+  font-family: "Alef-Regular", sans-serif;
+  font-size: 1rem;
+  font-weight: 400;
+  cursor: pointer;
+}
