+++ src/Profile/Signin/Signin.module.css
@@ -1,107 +1,5 @@
-/* General reset for elements */
-a,
-button,
-input,
-select,
-h1,
-h2,
-h3,
-h4,
-h5,
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  border: none;
-  text-decoration: none;
-  appearance: none;
-  background: none;
-  -webkit-font-smoothing: antialiased;
-}
 
-.signIn,
-.signIn * {
-  box-sizing: border-box;
-}
-
-/* .frame37 {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-evenly;
-  flex-shrink: 0;
-  position: relative;
-} */
-
-.signIn {
-  background: var(--white, #ffffff);
-  border-radius: 0.625rem;
-  /* padding: 1.0625rem 0.3125rem; */
-  display: flex;
-  flex-direction: column;
-  gap: 2.3125rem;
-  align-items: center;
-  justify-content: center;
-  /* height: 25.625rem; */
-  box-shadow: 0 0 0.375rem rgba(0, 0, 0, 0.25);
-  /*  */
-  width: 480px;
-  /* height: 621px; */
-
-  min-height: 621px;
-  background-color: white;
-  position: relative;
-  box-sizing: border-box;
-  margin: 50px auto;
-  padding: 20px;
-}
-
-.signInMain {
-  color: rgba(0, 0, 0, 0.67);
-  text-align: center;
-  font-family: "Alef-Regular", sans-serif;
-  font-size: 2rem;
-  line-height: 135.02%;
-  font-weight: 400;
-  position: relative;
-  width: 16.8125rem;
-  height: 2.625rem;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-}
-
-.frame36 {
-  margin: -0.875rem 0 0 0;
-  display: flex;
-  flex-direction: column;
-  gap: 0.575rem;
-  align-items: center;
-  justify-content: flex-start;
-  flex-shrink: 0;
-  position: relative;
-}
-
-.x {
-  flex-shrink: 0;
-  width: 1.5rem;
-  height: 1.5rem;
-  position: absolute;
-  right: 1.5rem;
-  top: 1.3125rem;
-  overflow: visible;
-}
-
-/* .SignInArea {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  position: relative;
-} */
-
-.frame33 {
+.SignInInputField {
   display: flex;
   flex-direction: column;
   gap: 1.0625rem;
@@ -117,6 +15,10 @@ h5,
   width: 20rem;
   height: 4.375rem;
   position: relative;
+}
+
+.signInTextBox:focus {
+  border: none;
 }
 
 .frame31 {
@@ -142,6 +44,10 @@ h5,
   /* top: calc(50% - 0.625rem); */
   height: 100%;
   width: 100%;
+}
+
+.placeholder {
+  outline: rgba(54, 54, 54, 0.649);
 }
 
 .nameLabel {
@@ -183,86 +89,6 @@ h5,
   justify-content: flex-start;
 }
 
-.frame34 {
-  display: flex;
-  flex-direction: column;
-  gap: 0.625rem;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  position: relative;
-}
-
-.socialLoginText {
-  color: rgba(0, 0, 0, 0.67);
-  text-align: center;
-  font-family: "Alef-Regular", sans-serif;
-  font-size: 1rem;
-  line-height: 135.02%;
-  font-weight: 400;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.frame32 {
-  display: flex;
-  flex-direction: row;
-  gap: 0.625rem;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  width: 19.3125rem;
-  position: relative;
-}
-
-.signUpText {
-  color: rgba(0, 0, 0, 0.67);
-  text-align: center;
-  font-family: "Alef-Regular", sans-serif;
-  font-size: 1rem;
-  line-height: 135.02%;
-  font-weight: 400;
-  position: relative;
-  align-self: stretch;
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.signUp {
-  color: rgba(13, 69, 179, 0.96);
-  text-align: center;
-  font-family: "Alef-Regular", sans-serif;
-  font-size: 1rem;
-  line-height: 135.02%;
-  font-weight: 400;
-  position: relative;
-  width: 4.125rem;
-  height: 2rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-}
-
-/*  */
-.modal {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.6);
-
-  /*  */
-  /* display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center; */
-}
 
 .loginModal {
   width: 480px;
