+++ src/Pureplate/Header/Header.module.css
@@ -11,6 +11,7 @@
     justify-content: flex-start;
     width: 100%;
     height: 5rem;
+    padding-left: 1.2rem;
     position: absolute;
     left: 0rem;
     top: 0rem;
@@ -24,13 +25,12 @@
   }
   
   .purePlateIcon {
-    width: 60px;
-    height: 60px;
+    width: 50px;
+    height: 50px;
     position: relative;
     align-items: center;
     object-fit: contain;
-    padding-right: 10px;
-    padding-top: 5px;
+    margin-right: 1rem;
   }
   
   @media (max-width: 600px) {
